Finished Dijskstras pathfinding (still needs functional testing)

+++ IDE Code/Buggy_Project/UltraSonicSensor.h
@@ -34,7 +34,7 @@ public:
   long duration, distance; // Duration used to calculate distance
 
 	// Return in cm the distance to an object
-	double getDistanceToNearestObject();
+	float getDistanceToNearestObject();
 
   bool isCollision();
 };
